+++ cli/src/delete/service.rs
@@ -6,7 +6,7 @@ use std::{
 use anyhow::{Context, Result};
 use clap::{Arg, ArgAction, ArgMatches, Command};
 use rustyline::{Editor, history::DefaultHistory};
-use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
+use termcolor::{ColorChoice, StandardStream, WriteColor};
 use toml::to_string_pretty as toml_to_string_pretty;
 
 use crate::{
@@ -133,9 +133,7 @@ impl CliCommand for ServiceCommand {
             )?;
 
             if !continue_delete {
-                stdout.set_color(ColorSpec::new().set_fg(Some(Color::Red)))?;
-                writeln!(stdout, "Deletion cancelled")?;
-                stdout.reset()?;
+                log_error!(stdout, "Deletion cancelled");
                 return Ok(());
             }
         }
@@ -244,9 +242,7 @@ impl CliCommand for ServiceCommand {
             .collect();
         write_rendered_templates(&rendered_templates, false, &mut stdout)?;
 
-        stdout.set_color(ColorSpec::new().set_fg(Some(Color::Green)))?;
-        writeln!(stdout, "{} deleted successfully!", service_name)?;
-        stdout.reset()?;
+        log_ok!(stdout, "{} deleted successfully!", service_name);
 
         Ok(())
     }